feat: load web only for dev
