+++ packages/vike-node/src/runtime/vike-handler.ts
@@ -2,7 +2,6 @@ import type { IncomingMessage, ServerResponse } from 'node:http'
 import compressMiddlewareFactory from '@universal-middleware/compress'
 import type { Get, RuntimeAdapter, UniversalHandler, UniversalMiddleware } from '@universal-middleware/core'
 import { renderPage as _renderPage } from 'vike/server'
-import { assert } from '../utils/assert.js'
 import { isVercel } from '../utils/isVercel.js'
 import { connectToWeb } from './adapters/connectToWeb.js'
 import { globalStore } from './globalStore.js'
@@ -76,6 +75,7 @@ export const renderPageHandler = ((options?) => async (request, context, runtime
   }
 
   if (globalStore.isDev) {
+    const { web } = await import('./dev/web.js')
     const handled = await web(request)
 
     if (handled) return handled
@@ -116,17 +116,3 @@ export const renderPageHandler = ((options?) => async (request, context, runtime
     headers: response.headers
   })
 }) satisfies Get<[options: VikeOptions], UniversalHandler>
-
-const web = connectToWeb(handleViteDevServer)
-
-function handleViteDevServer(req: IncomingMessage, res: ServerResponse): Promise<boolean> {
-  return new Promise<boolean>((resolve) => {
-    res.once('close', () => {
-      resolve(true)
-    })
-    assert(globalStore.viteDevServer)
-    globalStore.viteDevServer.middlewares(req, res, () => {
-      resolve(false)
-    })
-  })
-}
